Add function to auto-merge if resume fails
When resuming changesets, I hit a chunk of resumes that all required merging--however, auto-merging fixed the resume.  The command line help for "resume" describes a "--no-merge" option that disables automatic merging of conflicts, which implies that the "resume" command should be auto-merging.  For some reason, when the changesets were being resumed, they were not automerging.  

I added a function that resolves conflicts by auto-merging if the resume fails.