Do not set flow targets
As I was running the script on some streams, I kept getting CLIClientExceptions with FileSync errors.  The root cause is a defect in RTC that was fixed for 5.0.2: https://jazz.net/jazz/web/projects/Rational%20Team%20Concert#action=com.ibm.team.workitem.viewWorkItem&id=315286.  Unfortunately, my server is not 5.0.2, so the fix is to not set the flow target so that the CLI is not trying to track pending changes.  Before running the script, the flow target should be manually removed from the repository workspace.